requirements.txt

fake-useragent==1.5.1
loguru==0.7.2
pydantic-settings==2.6.1
undetected-playwright-patch==1.40.0.post1700587210000